refactor:refactor code in ItemTable and ItemDrawer
