+++ src/components/ItemDrawer.tsx
@@ -2,57 +2,12 @@ import { Drawer } from "@mui/material";
 import AddIcon from "@mui/icons-material/Add";
 import { useEffect, useState } from "react";
 import "./ItemDrawer.css";
-import {
-  DataGrid,
-  GridColDef,
-  GridRenderCellParams,
-  GridRowsProp,
-  GridValueFormatterParams,
-} from "@mui/x-data-grid";
+import { GridRowId, GridRowModel, GridRowsProp } from "@mui/x-data-grid";
 import { DataStore } from "aws-amplify";
-import TagsList from "./TagsList";
-// import { columns } from "./ItemTable";
-
-const columns: GridColDef[] = [
-  { field: "name", headerName: "Name", minWidth: 350 },
-  {
-    field: "value",
-    headerName: "Price",
-    minWidth: 200,
-    flex: 0.5,
-    valueFormatter: (params: GridValueFormatterParams) => {
-      const value = params.value as number;
-      return `$${value.toFixed(2)}`;
-    },
-  },
-  {
-    field: "dateAdded",
-    headerName: "Date Added",
-    minWidth: 200,
-    flex: 0.5,
-    valueFormatter: (params: GridValueFormatterParams) => {
-      const date = new Date(params.value as string);
-      return date.toLocaleDateString();
-    },
-  },
-  {
-    field: "cats",
-    headerName: "Categories",
-    minWidth: 350,
-    flex: 2,
-    renderCell: (params: GridRenderCellParams) => (
-      <TagsList categories={params.value as string[]} />
-    ),
-    sortComparator: (v1: string[], v2: string[]) => {
-      const v1Str = v1.join("");
-      const v2Str = v2.join("");
-      return v1Str.localeCompare(v2Str);
-    },
-  },
-];
+import Table from "./Table";
 
 const testRow = {
-  id: 2,
+  id: 2324,
   name: "New balance sneaker",
   dateAdded: "01 Jan 1970 00:00:00 GMT",
   value: 60,
@@ -60,11 +15,6 @@ const testRow = {
 };
 
 function ItemDrawer() {
-  const [paginationModel, setPaginationModel] = useState({
-    pageSize: 25,
-    page: 0,
-  });
-
   const [isDrawerOpen, setIsDrawerOpen] = useState(false);
 
   const [isDisabled, setIsDisabled] = useState(true);
@@ -92,6 +42,28 @@ function ItemDrawer() {
     }
     setRows([]);
   }
+
+  const handleDeleteClick = (id: GridRowId) => () => {
+    setRows(rows.filter((row) => row.id !== id));
+  };
+
+  // This function will only be called when user click edit and save button
+  // The newRow is the updated row, and the table will display the updated row
+  // TODO: modify this function to update the rows with the newRow
+  // suggestion: find the old row by id and update the field
+  const handleRowEditCommit = (newRow: GridRowModel) => {
+    let categories = newRow.cats;
+    if (typeof categories === "string") {
+      categories = categories.replace(/\s/g, "").split(",").sort();
+    }
+    // reformat the newRow to match the type of each column
+    const rowFormatter = {
+      ...newRow,
+      value: parseFloat(newRow.value),
+      cats: categories,
+    };
+    return rowFormatter;
+  };
 
   return (
     <>
@@ -142,24 +114,11 @@ function ItemDrawer() {
             </button>
           </div>
           <div className="data-table">
-            <DataGrid
+            <Table
+              tableHeight="54vh"
               rows={rows}
-              columns={columns}
-              getRowHeight={() => "auto"}
-              paginationModel={paginationModel}
-              onPaginationModelChange={setPaginationModel}
-              pageSizeOptions={[10, 25, 50]}
-              sx={{
-                "&.MuiDataGrid-root--densityCompact .MuiDataGrid-cell": {
-                  py: "8px",
-                },
-                "&.MuiDataGrid-root--densityStandard .MuiDataGrid-cell": {
-                  py: "15px",
-                },
-                "&.MuiDataGrid-root--densityComfortable .MuiDataGrid-cell": {
-                  py: "22px",
-                },
-              }}
+              handleDeleteClick={handleDeleteClick}
+              handleRowEditCommit={handleRowEditCommit}
             />
           </div>
           <div className="end-buttons">
