+++ src/components/ItemTable.tsx
@@ -1,52 +1,18 @@
-import {
-  GridRowModesModel,
-  GridRowModes,
-  DataGrid,
-  GridColDef,
-  GridActionsCellItem,
-  GridRowId,
-  GridRenderCellParams,
-  GridValueFormatterParams,
-  GridRowModel,
-} from "@mui/x-data-grid";
-import EditIcon from "@mui/icons-material/Edit";
-import DeleteIcon from "@mui/icons-material/DeleteOutlined";
-import SaveIcon from "@mui/icons-material/Save";
-import CancelIcon from "@mui/icons-material/Close";
-// import DeleteForeverIcon from "@mui/icons-material/DeleteForever";
-// import SaveIcon from "@mui/icons-material/Save";
+import { GridRowId, GridRowModel, GridRowsProp } from "@mui/x-data-grid";
 import { DataStore } from "aws-amplify";
 import { useEffect, useState } from "react";
 import { Category, Item, ItemCategory } from "../models";
-import "./TagStyle.css";
-import TagsList from "./TagsList";
+import Table from "./Table";
 
 export default function ItemTable() {
-  const [rows, setRows] = useState<Item[]>([]);
-  const [paginationModel, setPaginationModel] = useState({
-    pageSize: 25,
-    page: 0,
-  });
-  const [rowModesModel, setRowModesModel] = useState<GridRowModesModel>({});
-
-  const handleEditClick = (id: GridRowId) => () => {
-    console.log("editing id = ", id);
-    setRowModesModel({ ...rowModesModel, [id]: { mode: GridRowModes.Edit } });
-  };
-
-  const handleSaveClick = (id: GridRowId) => () => {
-    console.log("saving id = ", id);
-    setRowModesModel({ ...rowModesModel, [id]: { mode: GridRowModes.View } });
-  };
+  const [rows, setRows] = useState<GridRowsProp>([]);
 
   const handleDeleteClick = (id: GridRowId) => async () => {
     console.log("deleting id = ", id);
-    console.log("typeof id = ", typeof id);
     const rowToDelete = await DataStore.query(Item, (itemId) =>
       itemId.id.eq(id.toString())
     );
     console.log("toDelete =", rowToDelete);
-    console.log("rowToDelete.length =", rowToDelete.length);
 
     if (rowToDelete.length === 1) {
       try {
@@ -58,18 +24,6 @@ export default function ItemTable() {
     }
 
     setRows(rows.filter((row) => row.id !== id));
-  };
-
-  const handleCancelClick = (id: GridRowId) => () => {
-    console.log("canceling id = ", id);
-    setRowModesModel({
-      ...rowModesModel,
-      [id]: { mode: GridRowModes.View, ignoreModifications: true },
-    });
-  };
-
-  const handleRowModesModelChange = (newRowModesModel: GridRowModesModel) => {
-    setRowModesModel(newRowModesModel);
   };
 
   const handleRowEditCommit = async (newRow: GridRowModel) => {
@@ -98,6 +52,8 @@ export default function ItemTable() {
             Object.assign(updated, rowFormatterCopy);
           })
         );
+        console.log("Updated the Item = ", updatedItem);
+
         // update the Category
         if (editCats) {
           // Get all the existing category
@@ -143,6 +99,7 @@ export default function ItemTable() {
                   category: categoryToAdd[0],
                 });
                 await DataStore.save(categoryItem);
+                console.log("Added categoryItem = ", categoryItem);
               });
             }
 
@@ -151,7 +108,6 @@ export default function ItemTable() {
               .filter((category) => getOldCategoriesID.includes(category.id))
               .filter((category) => !categories.includes(category.name))
               .map((category) => category.id);
-            console.log("IdToDelete = ", IdToDelete);
             if (IdToDelete.length > 0) {
               IdToDelete.map(async (cID) => {
                 const categoryToDelete = await DataStore.query(
@@ -163,6 +119,7 @@ export default function ItemTable() {
                     ])
                 );
                 await DataStore.delete(categoryToDelete[0]);
+                console.log("Removed categoryItem = ", categoryToDelete[0]);
               });
             }
           }
@@ -171,98 +128,8 @@ export default function ItemTable() {
         console.log("Error! Cannot update the row ", error);
       }
     }
-
     return rowFormatter;
   };
-
-  const handleProcessRowUpdateError = (error: Error) => {
-    console.log("error = ", error);
-  };
-
-  const columns: GridColDef[] = [
-    { field: "name", headerName: "Name", minWidth: 350, editable: true },
-    {
-      field: "value",
-      headerName: "Price",
-      minWidth: 200,
-      flex: 0.5,
-      valueFormatter: (params: GridValueFormatterParams) => {
-        const value = params.value as number;
-        return `$${value.toFixed(2)}`;
-      },
-      editable: true,
-    },
-    {
-      field: "dateAdded",
-      headerName: "Date Added",
-      minWidth: 200,
-      flex: 0.5,
-      valueFormatter: (params: GridValueFormatterParams) => {
-        const date = new Date(params.value as string);
-        return date.toLocaleDateString();
-      },
-      editable: true,
-    },
-    {
-      field: "cats",
-      headerName: "Categories",
-      minWidth: 250,
-      flex: 2,
-      renderCell: (params: GridRenderCellParams) => (
-        <TagsList categories={params.value as string[]} />
-      ),
-      sortComparator: (v1: string[], v2: string[]) => {
-        const v1Str = v1.join("");
-        const v2Str = v2.join("");
-        return v1Str.localeCompare(v2Str);
-      },
-      editable: true,
-    },
-    {
-      field: "actions",
-      type: "actions",
-      headerName: "Actions",
-      minWidth: 200,
-      flex: 1,
-      cellClassName: "actions",
-      getActions: ({ id }) => {
-        const isInEditMode = rowModesModel[id]?.mode === GridRowModes.Edit;
-
-        if (isInEditMode) {
-          return [
-            <GridActionsCellItem
-              icon={<SaveIcon />}
-              label="Save"
-              onClick={handleSaveClick(id)}
-            />,
-            <GridActionsCellItem
-              icon={<CancelIcon />}
-              label="Cancel"
-              className="textPrimary"
-              onClick={handleCancelClick(id)}
-              color="inherit"
-            />,
-          ];
-        }
-
-        return [
-          <GridActionsCellItem
-            icon={<EditIcon />}
-            label="Edit"
-            className="textPrimary"
-            onClick={handleEditClick(id)}
-            color="inherit"
-          />,
-          <GridActionsCellItem
-            icon={<DeleteIcon />}
-            label="Delete"
-            onClick={handleDeleteClick(id)}
-            color="inherit"
-          />,
-        ];
-      },
-    },
-  ];
 
   useEffect(() => {
     const fetchData = async () => {
@@ -290,34 +157,15 @@ export default function ItemTable() {
         console.log("Error retrieving posts", error);
       }
     };
-
     fetchData();
   }, []);
 
   return (
-    <div style={{ height: "70vh" }}>
-      <DataGrid
-        rows={rows}
-        columns={columns}
-        getRowHeight={() => "auto"}
-        editMode="row"
-        rowModesModel={rowModesModel}
-        onRowModesModelChange={handleRowModesModelChange}
-        processRowUpdate={handleRowEditCommit}
-        onProcessRowUpdateError={handleProcessRowUpdateError}
-        paginationModel={paginationModel}
-        onPaginationModelChange={setPaginationModel}
-        pageSizeOptions={[10, 25, 50]}
-        sx={{
-          "&.MuiDataGrid-root--densityCompact .MuiDataGrid-cell": { py: "8px" },
-          "&.MuiDataGrid-root--densityStandard .MuiDataGrid-cell": {
-            py: "15px",
-          },
-          "&.MuiDataGrid-root--densityComfortable .MuiDataGrid-cell": {
-            py: "22px",
-          },
-        }}
-      />
-    </div>
+    <Table
+      tableHeight="70vh"
+      rows={rows}
+      handleDeleteClick={handleDeleteClick}
+      handleRowEditCommit={handleRowEditCommit}
+    />
   );
 }
